All Bugs Fixed Scouting App Completed

+++ scoutingapp/src/main/java/org/usfirst/frc/team25/scouting/data/models/TeleOp.java
@@ -60,7 +60,7 @@ public class TeleOp {
        this.climbTime = climbTime;
     }
 
-    public int getClimbTime(){return climbTime;}
+    public long getClimbTime(){return climbTime;}
     public int getRobotCargoPickedUp() {
         return robotCargoPickedUp;
     }
